+++ NKATM/NKATM.SocialFund/NKATM.SocialFund.Application/Features/SocialResults/Handlers/Commands/DeleteSocialResultCommandHandler.cs
@@ -1,7 +1,9 @@
 ﻿using AutoMapper;
 using MediatR;
+using NKATM.SocialFund.Application.Exceptions;
 using NKATM.SocialFund.Application.Features.SocialResults.Commands;
 using NKATM.SocialFund.Application.Persistance.Contracts;
+using NKATM.SocialFund.Domain;
 using System;
 using System.Collections.Generic;
 using System.Text;
@@ -24,6 +26,12 @@ namespace NKATM.SocialFund.Application.Features.SocialResults.Handlers.Commands
         public async Task<Unit> Handle(DeleteSocialResultCommand request, CancellationToken cancellationToken)
         {
             var socialResult = await repository.GetAsync(request.Id);
+
+            if (socialResult == null)
+            {
+                throw new NotFoundException(nameof(SocialResult), request.Id);
+            }
+
             await repository.DeleteAsync(socialResult);
 
             return Unit.Value;
